@@ -2,19 +2,19 @@
 
 declare(strict_types=1);
 
-namespace Jahudka\ComponentEvents\Bridges\Symfony;
+namespace Jahudka\ComponentEvents\Bridges\NextrasORM;
 
 use Nette\DI\ContainerBuilder;
 use Nette\DI\Definitions\Statement;
 use Jahudka\ComponentEvents\IBridge;
 use Jahudka\ComponentEvents\IAnalyser;
-use Symfony\Component\EventDispatcher\EventDispatcher;
+use Nextras\Orm\Model\Model;
 
 
 class Bridge implements IBridge {
 
     public function detectPresence(ContainerBuilder $builder) : bool {
-        return !empty($builder->findByType(EventDispatcher::class));
+        return !empty($builder->findByType(Model::class));
     }
 
     public function createAnalyser() : IAnalyser {
@@ -22,30 +22,28 @@ class Bridge implements IBridge {
     }
 
     public function createRelayFactory(array $listeners) : Statement {
-        $eventMap = $this->createEventMap($listeners);
-
         return new Statement(
             Relay::class,
             [
-                'eventMap' => $eventMap,
+                'eventMap' => $this->createEventMap($listeners),
             ]
         );
     }
 
     private function createEventMap(array $listenerMap) : array {
-        $eventMap = [];
+        $map = [];
 
         foreach ($listenerMap as $presenter => $components) {
             foreach ($components as $component => $events) {
-                foreach ($events as $event => $listeners) {
-                    foreach ($listeners as $listener) {
-                        $eventMap[$presenter][$event][$listener[1] ?? 0][$component][] = $listener[0];
+                foreach ($events as $event => $entities) {
+                    foreach ($entities as $entity) {
+                        $map[$presenter][$entity][$event][] = $component;
                     }
                 }
             }
         }
 
-        return $eventMap;
+        return $map;
     }
 
 }
