@@ -6,11 +6,12 @@ namespace Jahudka\ComponentEvents\Bridges\Symfony;
 
 use Jahudka\ComponentEvents\IAnalyser;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use ReflectionClass;
 
 
 class Analyser implements IAnalyser {
 
-    public function analyse(\ReflectionClass $component) : ?array {
+    public function analyse(ReflectionClass $component) : ?array {
         if (!$component->implementsInterface(EventSubscriberInterface::class)) {
             return null;
         }
@@ -20,7 +21,7 @@ class Analyser implements IAnalyser {
         return array_map(function($listeners) {
             if (is_string($listeners)) {
                 return [[$listeners]];
-            } else if (is_string($listeners[0])) {
+            } else if ($listeners && is_string($listeners[0])) {
                 return [$listeners];
             } else {
                 return $listeners;
