@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Jahudka\ComponentEvents;
 
+use ReflectionClass;
+
 
 interface IAnalyser {
 
-    public function analyse(\ReflectionClass $component) : ?array;
+    public function analyse(ReflectionClass $component) : ?array;
 
 }
